✨ Generate column definitions from schema
