+++ packages/ui/react-ui-table/package.json
@@ -41,6 +41,7 @@
     "@phosphor-icons/react": "^2.0.5",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
+    "fast-check": "^3.15.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "vite": "^5.1.5",
